Working on a bug fix. The default column collation of "contents" within the article table cannot handle umlauts. Additionally, the code I wrote, encodes the contents of an article to base64 first clientside when composing, the server subsequently decodes the contents and inserts them raw into the table. Because the contents may contain umlauts, and the databases collation isn't compatible, the umlauts can't be translated and thus, are lost in the process.
Now there are three possible solutions: alter the collation to support umlauts, keep the contents base64 encoded in the table, or both... which would be ideal.

The issues faced during troubleshooting: I have removed the part of the code that decodes the incoming changes from the composer page, decode the relevant data within the jsp files and display it that way.